Separate boxes into separate files ... not working yet.
