+++ src/mpeg4/boxes/appleAdditionalInfoBox.ts
@@ -1,0 +1,69 @@
+import FullBox from "./fullBox";
+import IsoHandlerBox from "./isoHandlerBox";
+import Mpeg4BoxHeader from "../mpeg4BoxHeader";
+import {ByteVector, StringType} from "../../byteVector";
+import {File} from "../../file";
+import {Mpeg4BoxClassType} from "../mpeg4BoxClassType";
+import {StringUtils} from "../../utils";
+
+/**
+ *  This class extends @see FullBox to provide an implementation of an Apple AdditionalInfoBox.
+ */
+export default class AppleAdditionalInfoBox extends FullBox {
+    /**
+     * Private constructor to force construction via static functions.
+     */
+    private constructor() {
+        super();
+    }
+
+    /**
+     * Constructs and initializes a new instance of @see AppleAdditionalInfoBox with a provided header
+     * and handler by reading the contents from a specified file.
+     * @param header A @see Mpeg4BoxHeader object containing the header to use for the new instance.
+     * @param file A @see File object to read the contents of the box from.
+     * @param handler A @see IsoHandlerBox object containing the handler that applies to the new instance.
+     * @returns A new instance of @see AppleAdditionalInfoBox.
+     */
+    public static fromHeaderFileAndHandler(
+        header: Mpeg4BoxHeader,
+        file: File,
+        handler: IsoHandlerBox
+    ): AppleAdditionalInfoBox {
+        const instance: AppleAdditionalInfoBox = new AppleAdditionalInfoBox();
+        instance.initializeFromHeaderFileAndHandler(header, file, handler);
+        instance.data = file.readBlock(instance.dataSize > 0 ? instance.dataSize : 0);
+
+        return instance;
+    }
+
+    /** @inheritDoc */
+    public get boxClassType(): Mpeg4BoxClassType { return Mpeg4BoxClassType.AppleAdditionalInfoBox; }
+
+    /**
+     * Gets the text contained in the current instance.
+     */
+    public get text(): string {
+        return StringUtils.trimStart(this.data.toString(StringType.Latin1), "\0");
+    }
+    /**
+     * Sets the text contained in the current instance.
+     */
+    public set text(v: string) {
+        this.data = ByteVector.fromString(v, StringType.Latin1);
+    }
+
+    /**
+     * Constructs and initializes a new instance of @see FullBox with a provided header, version, and flags.
+     * @param type A @see Mpeg4BoxHeader object containing the header to use for the new instance.
+     * @param version A value containing the version of the new instance.
+     * @param flags A value containing the flags for the new instance.
+     * @returns A new instance of @see FullBox.
+     */
+    public static fromTypeVersionAndFlags(type: ByteVector, version: number, flags: number): AppleAdditionalInfoBox {
+        const instance: AppleAdditionalInfoBox = new AppleAdditionalInfoBox();
+        instance.initializeFromTypeVersionAndFlags(type, version, flags);
+
+        return instance;
+    }
+}
