+++ src/mpeg4/mpeg4BoxHeader.ts
@@ -1,7 +1,7 @@
+import Mpeg4Box from "./boxes/mpeg4Box";
 import Mpeg4BoxType from "./mpeg4BoxType";
 import { ByteVector, StringType } from "../byteVector";
 import { File } from "../file";
-import { Mpeg4Box } from "./mpeg4Boxes";
 import {Guards, NumberUtils} from "../utils";
 
 /**
